Add files via upload
Did the minimal tasks (sorted the dropdown, not creating a marker each time an airport is selected,
	filled out the rest of the data and foxthough I later hid the displayed information).
Added a tooltip with the location to the dropdown and renamed the label.
Added a corresponding dropdown for the location.
The two dropdowns stay in sync and are sufficient for showing the Airport code and location (city, state),
	which I assume is the most important info.
To preserve screen real estate, I hid the table of information that was showing and
moved it to a tooltip on the marker. (The tooltip cleans up the full name.)
Added a button to hide/show the current marker.
Replaced the project instructions with instructions for the updated page.

The Maps API does not have a built-in context menu.  There were some 3rd party libraries
that do them.  If I had more time, I would have hooked that up and added a context menu
on right click for the markers and had a delete marker item in it.  A